chore: reformat structure, add prettier, add hooks
