+++ .prettierignore
@@ -1,0 +1,5 @@
+# ignore artifacts
+build
+coverage
+
+**/*.yaml
