+++ .storybook/preview.ts
@@ -1,17 +1,17 @@
-import type { Preview } from "@storybook/react";
+import type { Preview } from '@storybook/react';
 
 import '../src/index.css';
 
 const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: "^on[A-Z].*" },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
+    parameters: {
+        actions: { argTypesRegex: '^on[A-Z].*' },
+        controls: {
+            matchers: {
+                color: /(background|color)$/i,
+                date: /Date$/,
+            },
+        },
     },
-  },
 };
 
 export default preview;
