+++ package.json
@@ -1,66 +1,75 @@
 {
-  "name": "new",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.53",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest",
-      "plugin:storybook/recommended"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@storybook/addon-essentials": "^7.4.3",
-    "@storybook/addon-interactions": "^7.4.3",
-    "@storybook/addon-links": "^7.4.3",
-    "@storybook/addon-onboarding": "^1.0.8",
-    "@storybook/addon-styling": "^1.3.7",
-    "@storybook/blocks": "^7.4.3",
-    "@storybook/preset-create-react-app": "^7.4.3",
-    "@storybook/react": "^7.4.3",
-    "@storybook/react-webpack5": "^7.4.3",
-    "@storybook/testing-library": "^0.2.1",
-    "autoprefixer": "^10.4.16",
-    "babel-plugin-named-exports-order": "^0.0.2",
-    "eslint-plugin-storybook": "^0.6.14",
-    "postcss": "^8.4.30",
-    "prop-types": "^15.8.1",
-    "storybook": "^7.4.3",
-    "tailwindcss": "^3.3.3",
-    "webpack": "^5.88.2"
-  }
+    "name": "new",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "@testing-library/jest-dom": "^5.17.0",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^13.5.0",
+        "@types/jest": "^27.5.2",
+        "@types/node": "^16.18.53",
+        "@types/react": "^18.2.22",
+        "@types/react-dom": "^18.2.7",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-scripts": "5.0.1",
+        "typescript": "^4.9.5",
+        "web-vitals": "^2.1.4"
+    },
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject",
+        "storybook": "storybook dev -p 6006",
+        "build-storybook": "storybook build",
+        "prettier": "prettier --log-level warn --write \"src/**/*.{tsx,ts}\"",
+        "prettier-check": "prettier --log-level warn --check \"src/**/*.{tsx,ts}\"",
+        "prepare": "husky install"
+    },
+    "lint-staged": {
+        "**/*": "prettier --write --ignore-unknown"
+    },
+    "eslintConfig": {
+        "extends": [
+            "react-app",
+            "react-app/jest",
+            "plugin:storybook/recommended"
+        ]
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "devDependencies": {
+        "@storybook/addon-essentials": "^7.4.3",
+        "@storybook/addon-interactions": "^7.4.3",
+        "@storybook/addon-links": "^7.4.3",
+        "@storybook/addon-onboarding": "^1.0.8",
+        "@storybook/addon-styling": "^1.3.7",
+        "@storybook/blocks": "^7.4.3",
+        "@storybook/preset-create-react-app": "^7.4.3",
+        "@storybook/react": "^7.4.3",
+        "@storybook/react-webpack5": "^7.4.3",
+        "@storybook/testing-library": "^0.2.1",
+        "autoprefixer": "^10.4.16",
+        "babel-plugin-named-exports-order": "^0.0.2",
+        "eslint-plugin-storybook": "^0.6.14",
+        "husky": "^8.0.3",
+        "lint-staged": "^14.0.1",
+        "postcss": "^8.4.30",
+        "prettier": "3.0.3",
+        "prop-types": "^15.8.1",
+        "storybook": "^7.4.3",
+        "tailwindcss": "^3.3.3",
+        "webpack": "^5.88.2"
+    }
 }
