+++ src/components/page/page.css
@@ -1,0 +1,69 @@
+.storybook-page {
+    font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
+    font-size: 14px;
+    line-height: 24px;
+    padding: 48px 20px;
+    margin: 0 auto;
+    max-width: 600px;
+    color: #333;
+}
+
+.storybook-page h2 {
+    font-weight: 700;
+    font-size: 32px;
+    line-height: 1;
+    margin: 0 0 4px;
+    display: inline-block;
+    vertical-align: top;
+}
+
+.storybook-page p {
+    margin: 1em 0;
+}
+
+.storybook-page a {
+    text-decoration: none;
+    color: #1ea7fd;
+}
+
+.storybook-page ul {
+    padding-left: 30px;
+    margin: 1em 0;
+}
+
+.storybook-page li {
+    margin-bottom: 8px;
+}
+
+.storybook-page .tip {
+    display: inline-block;
+    border-radius: 1em;
+    font-size: 11px;
+    line-height: 12px;
+    font-weight: 700;
+    background: #e7fdd8;
+    color: #66bf3c;
+    padding: 4px 12px;
+    margin-right: 10px;
+    vertical-align: top;
+}
+
+.storybook-page .tip-wrapper {
+    font-size: 13px;
+    line-height: 20px;
+    margin-top: 40px;
+    margin-bottom: 40px;
+}
+
+.storybook-page .tip-wrapper svg {
+    display: inline-block;
+    height: 12px;
+    width: 12px;
+    margin-right: 4px;
+    vertical-align: top;
+    margin-top: 3px;
+}
+
+.storybook-page .tip-wrapper svg path {
+    fill: #1ea7fd;
+}
