+++ tailwind.config.js
@@ -1,9 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./src/**/*.{ts,tsx,mdx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
-
+    content: ['./src/**/*.{ts,tsx,mdx}'],
+    theme: {
+        extend: {},
+    },
+    plugins: [],
+};
